+++ browser_test/wed_test.js
@@ -2313,6 +2313,83 @@ describe("wed", function () {
 
             it("is able to start", function () { });
         });
+
+        describe("attribute errors without attributes being " +
+                 "shown (due to wed options)",
+                 function () {
+            before(function() {
+                var new_options = $.extend(true, {}, option_stack[0]);
+                new_options.mode.options.hide_attributes = true;
+                option_stack.unshift(new_options);
+            });
+
+            after(function () {
+                option_stack.shift();
+            });
+
+            it("is able to start", function (done) {
+                editor.whenCondition("first-validation-complete",
+                                     function () {
+                    done();
+                });
+            });
+        });
+
+        describe("attribute errors without attributes being " +
+                 "shown (because of the label visibility level)",
+                 function () {
+            beforeEach(function (done) {
+                editor.whenCondition("first-validation-complete",
+                                     function () {
+                    done();
+                });
+            });
+            it("the attributes error are not linked", function () {
+                while (editor._current_label_level)
+                    editor.decreaseLabelVisiblityLevel();
+
+                var errors = editor._validation_errors;
+                var attribute_errors = [];
+                var $items = editor.$error_list.children("li");
+                var cases = 0;
+                for (var i = 0, error; (error = errors[i]); ++i) {
+                    if (error.node.nodeType === Node.ATTRIBUTE_NODE) {
+                        var item = $items[i];
+                        assert.isTrue(
+                            item.getElementsByTagName("a").length === 0,
+                            "there should be no link in the item");
+                        assert.equal(
+                            item.title,
+                            "This error belongs to an attribute " +
+                                "which is not currently displayed.",
+                            "the item should have the right title");
+                        cases++;
+                    }
+                }
+
+                assert.equal(cases, 2);
+            });
+        });
+
+        describe("does not have completion menu", function () {
+            it("when the caret is in an attribute that takes " +
+               "completions but the attribute is not visible",
+               function () {
+                // Reduce visibility to 0 so that no attribute is
+                // visible.
+                while (editor._current_label_level)
+                    editor.decreaseLabelVisiblityLevel();
+                var p = editor.gui_root.querySelectorAll(".body>.p")[9];
+                var attr_vals = getAttributeValuesFor(p);
+                editor.setGUICaret(attr_vals[0].firstChild, 0);
+                // This is an arbitrary menu item we check for.
+
+                var menu = editor.my_window.document
+                    .getElementsByClassName("wed-context-menu")[0];
+                assert.isUndefined(menu, "the menu should not exist");
+            });
+        });
+
     });
 
     describe("(not state-sensitive)", function () {
@@ -2353,7 +2430,8 @@ describe("wed", function () {
             assert.isDefined(menu, "the menu should exist");
             var items = menu.querySelectorAll("li>a");
             var found = false;
-            for(var i = 0, item; !found && (item = items[i]) !== undefined;
+            for(var i = 0, item; !found &&
+                        (item = items[i]) !== undefined;
                 ++i) {
                 found = pattern.test(item.textContent.trim());
             }
@@ -2396,11 +2474,11 @@ describe("wed", function () {
 
         describe("setNavigationList", function () {
             it("makes the navigation list appear", function () {
-                assert.equal(editor._$navigation_panel.css("display"), "none",
-                            "the list is not displayed");
+                assert.equal(editor._$navigation_panel.css("display"),
+                             "none", "the list is not displayed");
                 editor.setNavigationList("foo");
-                assert.equal(editor._$navigation_panel.css("display"), "block",
-                            "the list is displayed");
+                assert.equal(editor._$navigation_panel.css("display"),
+                             "block", "the list is displayed");
             });
         });
 
@@ -2949,6 +3027,108 @@ describe("wed", function () {
                          gui_root.getElementsByClassName("wed-validation-error")
                          .length,
                          "the number of markers should be the same");
+        });
+
+        it("shows validation errors for inline elements in a correct " +
+           "position", function () {
+            editor.validator._validateUpTo(editor.data_root, -1);
+
+            var p = ps[11];
+            var data_p = editor.toDataNode(p);
+            var data_monogr = data_p.childNodes[0];
+            var monogr = $.data(data_monogr, "wed_mirror_node");
+            assert.equal(data_monogr.tagName, "monogr");
+
+            var errors = editor._validation_errors;
+            var p_error, p_error_ix;
+            var monogr_error, monogr_error_ix;
+            for (var i = 0, error; (error = errors[i]); ++i) {
+                if (!p_error && error.node === data_p) {
+                    p_error = error;
+                    p_error_ix = i;
+                }
+
+                if (!monogr_error && error.node === data_monogr) {
+                    monogr_error = error;
+                    monogr_error_ix = i;
+                }
+            }
+
+            // Make sure we found our errors.
+            assert.isDefined(p_error, "no error for our paragraph");
+            assert.isDefined(monogr_error, "no error for our monogr");
+
+            // Find the corresponding markers
+            var $markers = editor._$error_layer.
+                children(".wed-validation-error");
+            var p_marker = $markers[p_error_ix];
+            var monogr_marker = $markers[monogr_error_ix];
+            assert.isDefined(p_marker,
+                             "should have an error for our paragraph");
+            assert.isDefined(monogr_marker,
+                             "should have an error for our monogr");
+
+            var p_marker_rect = p_marker.getBoundingClientRect();
+
+            // The p_marker should appear to the right of the start
+            // label for the paragraph and overlap with the start
+            // label for monogr.
+            var p_start_label = firstGUI(p);
+            assert.isTrue(
+                p_start_label.classList.contains("__start_label"),
+                "should should have a start label for the paragraph");
+            var p_start_label_rect =
+                p_start_label.getBoundingClientRect();
+            assert.isTrue(p_marker_rect.left >= p_start_label_rect.right,
+                         "the paragraph error marker should be to " +
+                          "the right of the start label for the " +
+                          "paragraph");
+            assert.isTrue(Math.abs(p_marker_rect.bottom -
+                                   p_start_label_rect.bottom) <= 5,
+                         "the paragraph error marker should have " +
+                          "a bottom which is within 5 pixels of the " +
+                          "bottom of the start label for the paragraph");
+            assert.isTrue(Math.abs(p_marker_rect.top -
+                                   p_start_label_rect.top) <= 5,
+                         "the paragraph error marker should have " +
+                          "a top which is within 5 pixels of the " +
+                          "top of the start label for the paragraph");
+
+            var monogr_start_label = firstGUI(monogr);
+            assert.isTrue(
+                monogr_start_label.classList.contains("__start_label"),
+                "should should have a start label for the paragraph");
+            var monogr_start_label_rect =
+                monogr_start_label.getBoundingClientRect();
+            assert.isTrue(Math.abs(p_marker_rect.left -
+                                   monogr_start_label_rect.left) <= 5,
+                         "the paragraph error marker have a left side " +
+                          "within 5 pixels of the left side of the " +
+                          "start label for the monogr");
+
+
+            // The monogr_marker should be to the right of the
+            // monogr_start_label.
+
+            var monogr_marker_rect =
+                monogr_marker.getBoundingClientRect();
+
+            assert.isTrue(monogr_marker_rect.left >=
+                          monogr_start_label_rect.right,
+                          "the monogr error marker should be to " +
+                          "the right of the start label for the " +
+                          "monogr");
+            monogr_marker.scrollIntoView();
+            assert.isTrue(Math.abs(monogr_marker_rect.bottom -
+                                   monogr_start_label_rect.bottom) <= 5,
+                         "the monogr error marker should have " +
+                          "a bottom which is within 5 pixels of the " +
+                          "bottom of the start label for the monogr");
+            assert.isTrue(Math.abs(monogr_marker_rect.top -
+                                   monogr_start_label_rect.top) <= 5,
+                         "the monogr error marker should have " +
+                          "a top which is within 5 pixels of the " +
+                          "top of the start label for the monogr");
         });
 
         describe("the location bar", function () {