+++ selenium_test/steps/completion.py
@@ -5,25 +5,25 @@ from nose.tools import assert_equal, assert_true
 step_matcher("re")
 
 
-@then("^a completion menu is visible$")
+@then("a completion menu is visible")
 def step_impl(context):
     context.util.find_element((By.CLASS_NAME, "wed-completion-menu"))
 
 
-@then("^a completion menu is not visible$")
+@then("a completion menu is not visible")
 def step_impl(context):
     context.util.wait_until_not(
         lambda driver: driver.find_element_by_class_name(
             "wed-completion-menu"))
 
 
-@then("^the first item of the completion menu is focused$")
+@then("the first item of the completion menu is focused")
 def step_impl(context):
     context.util.find_element((By.CSS_SELECTOR,
                                ".wed-completion-menu li a:focus"))
 
 
-@then("^the completion text is inserted$")
+@then("the completion text is inserted")
 def step_impl(context):
     driver = context.driver
 
@@ -35,8 +35,8 @@ def step_impl(context):
     assert_equal(text, "initial")
 
 
-@then('^the completion menu has only one option named "initial" and the '
-      'prefix "i" is in bold$')
+@then('the completion menu has only one option named "initial" and the '
+      'prefix "i" is in bold')
 def step_impl(context):
     driver = context.driver