+++ selenium_test/steps/typeahead_popup.py
@@ -1,13 +1,14 @@
 from selenium.webdriver.common.action_chains import ActionChains
 from selenium.webdriver.common.by import By
 from nose.tools import assert_equal, assert_true
+from selenic.util import Result, Condition
 
 step_matcher("re")
 
 
-@when(ur"^(?P<step>the user opens the typeahead popup"
-      ur"(?P<where> at the end of the title text)?)$")
-@given("^(?P<step>that a typeahead popup is open)$")
+@when(ur"(?P<step>the user opens the typeahead popup"
+      ur"(?P<where> at the end of the title text)?)")
+@given("(?P<step>that a typeahead popup is open)")
 def step_impl(context, step, where=None):
     driver = context.driver
     util = context.util
@@ -44,10 +45,21 @@ def step_impl(context, step, where=None):
     When the user clicks the choice named "Test typeahead"
     """)
 
-    util.find_element((By.CLASS_NAME, "wed-typeahead-popup"))
+    def check(driver):
+        ret = driver.execute_script("""
+        var input = document.querySelector(
+          '.wed-typeahead-popup input.tt-input');
+        if (!input)
+          return [false, "cannot find input element"];
+        return [document.activeElement === input, "input not focused"];
+        """)
+        return Result(ret[0], ret[1])
+
+    result = Condition(util, check).wait()
+    assert_true(result, result.payload)
 
 
-@then("^the typeahead popup is not visible$")
+@then("the typeahead popup is not visible")
 def step_impl(context):
     util = context.util
 
@@ -56,7 +68,7 @@ def step_impl(context):
         len(driver.find_elements_by_class_name("wed-typeahead-popup")) == 0)
 
 
-@then("^the typeahead popup's action (?P<is_>is|is not) performed$")
+@then("the typeahead popup's action (?P<is_>is|is not) performed")
 def step_impl(context, is_):
     util = context.util
 
@@ -65,7 +77,7 @@ def step_impl(context, is_):
     assert_equal(util.get_text_excluding_children(element), expected)
 
 
-@when(ur"^the user clicks the first typeahead choice$")
+@when(ur"the user clicks the first typeahead choice")
 def step_impl(context):
     util = context.util
 
@@ -75,7 +87,7 @@ def step_impl(context):
     element.click()
 
 
-@when(ur"^the user clicks outside the typeahead$")
+@when(ur"the user clicks outside the typeahead")
 def step_impl(context):
     util = context.util
 
@@ -85,7 +97,7 @@ def step_impl(context):
         .perform()
 
 
-@then(ur"^the typeahead popup's choice list has a vertical scrollbar$")
+@then(ur"the typeahead popup's choice list has a vertical scrollbar")
 def step_impl(context):
     def check(driver):
         return driver.execute_script("""
@@ -104,7 +116,7 @@ def step_impl(context):
     context.util.wait(check)
 
 
-@then(ur"^the typeahead popup is visible and completely inside the window$")
+@then(ur"the typeahead popup is visible and completely inside the window")
 def step_impl(context):
     util = context.util
 
@@ -113,7 +125,7 @@ def step_impl(context):
                 "menu is completely visible")
 
 
-@then(ur'^the typeahead popup overflows the editor pane$')
+@then(ur'the typeahead popup overflows the editor pane')
 def step_impl(context):
     driver = context.driver
 
@@ -127,7 +139,7 @@ def step_impl(context):
     """), "the typeahead should overflow the editor pane")
 
 
-@when(ur"^the user clicks the last visible completion$")
+@when(ur"the user clicks the last visible completion")
 def step_impl(context):
     driver = context.driver
 
@@ -143,19 +155,19 @@ def step_impl(context):
         .perform()
 
 
-@then(ur"^dump caret position$")
+@then(ur"dump caret position")
 def step_impl(context):
     driver = context.driver
-    print driver.execute_script("""
+    print(driver.execute_script("""
     var caret = wed_editor.getGUICaret(true);
     if (!caret)
       return "no caret!"
     return [caret.node.innerHTML, caret.offset];
-    """)
-    print
+    """))
+    print("")
 
 
-@then(ur"^the typeahead popup shows suggestions$")
+@then(ur"the typeahead popup shows suggestions")
 def step_impl(context):
     util = context.util